Move target related code to LLVMTarget.
