+++ src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITEngine.h
@@ -54,8 +54,6 @@ class LLVMJITEngineBuilder {
   std::unique_ptr<LLVMJITEngine> build();
 
  private:
-  llvm::TargetMachine* buildTargetMachine();
-
   void dumpASM(LLVMJITEngine& engine);
 
   LLVMJITModule& module_;
